% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Bluejay_v0.1.R
\docType{package}
\name{DrugResponse}
\alias{DrugResponse}
\alias{DrugResponse-package}
\title{Bluejay}
\description{
R package use for classify ovarian cancer patients drug response by CA125 profile
}
\details{
Required packages: ggplot2, grid
}

